adding authorization to user for edit create delete

+++ routes/campgrounds.js
@@ -1,20 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const catchAsync = require('../utils/catchAsync.js');
-const ExpressError = require('../utils/ExpressError.js');
 const Campground = require('../models/campground.js');
-const { campSchema } = require('../joischema.js');
-const { isLoggedIn } = require('../middlewares');
-
-const validateCamp = (req, res, next) => {
-    const { error } = campSchema.validate(req.body);
-    if (error) {
-        const msg = error.details.map(e => e.message).join(',');
-        throw new ExpressError(msg, 400);
-    } else {
-        next();
-    }
-};
+const { isLoggedIn, isAuthor, validateCamp } = require('../middlewares');
 
 router.get('/', catchAsync(async (req, res) => {
     const campgrounds = await Campground.find();
@@ -26,7 +14,7 @@ router.get('/new', isLoggedIn, catchAsync(async (req, res) => {
 }));
 
 router.get('/:id', catchAsync(async (req, res) => {
-    const camp = await Campground.findById(req.params.id).populate('reviews');
+    const camp = await (await Campground.findById(req.params.id).populate('reviews').populate('author'));
     if (!camp) {
         req.flash('error', 'Cannot find that Campground! :(');
         return res.redirect('/campgrounds');
@@ -37,13 +25,15 @@ router.get('/:id', catchAsync(async (req, res) => {
 router.post('/', isLoggedIn, validateCamp, catchAsync(async (req, res) => {
     //if (!req.body.campground) throw new ExpressError('Invalid Campground Data', 400);
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id;
     await campground.save();
     req.flash('success', 'Successfully made a new Campground!');
     res.redirect(`/campgrounds/${campground._id}`);
 }));
 
-router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
     if (!campground) {
         req.flash('error', 'Cannot find that Campground! :(');
         return res.redirect('/campgrounds');
@@ -51,7 +41,7 @@ router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
     res.render('campgrounds/edit', { campground });
 }));
 
-router.put('/:id', isLoggedIn, validateCamp, catchAsync(async (req, res) => {
+router.put('/:id', isLoggedIn, isAuthor, validateCamp, catchAsync(async (req, res) => {
     const { id } = req.params;
     const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
     req.flash('success', 'Successfully updated the Campground!');
@@ -59,8 +49,13 @@ router.put('/:id', isLoggedIn, validateCamp, catchAsync(async (req, res) => {
     // res.send("it worked");
 }));
 
-router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
-    const { id } = req.params;;
+router.delete('/:id', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You do not have permissions!');
+        return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted the Campground!');
     res.redirect('/campgrounds')
